+++ eslint.config.js
@@ -1,5 +1,5 @@
 /*!
- * Copyright 2024 Digital Bazaar, Inc.
+ * Copyright 2022 - 2026 Digital Bazaar, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,17 +16,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-module.exports = {
-  root: true,
-  env: {
-    node: true
-  },
-  extends: [
-    'digitalbazaar',
-    'digitalbazaar/jsdoc',
-    'digitalbazaar/module'
-  ],
-  rules: {
-    'unicorn/prefer-node-protocol': 'error'
-  }
-};
+import config from '@digitalbazaar/eslint-config/node-recommended';
+
+export default [
+  ...config
+];